# pkg/server/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Marionette API
  version: 1.0.0
  description: |-
    The public Marionette API: create sessions, run tasks on them, answer the
    agent's permission requests, and stream the output.

    Authenticate with an API key as `Authorization: Bearer mk_...`. Each route
    requires a scope, listed in its security requirement. The two WebSocket routes
    also accept the key as a `?token=` query parameter, because a browser cannot
    set headers on a WebSocket handshake.

    List endpoints are cursor-paginated: pass `next_cursor` from one response as
    `cursor` on the next request, and stop when `has_more` is false. Repeated
    filters such as `status` are sent as repeated keys (`?status=a&status=b`),
    not as `status[]`.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/events:
    get:
      tags:
        - Streaming
      summary: Stream server events over WebSocket
      description: Emits permission requests and session, task and runner state changes.
      operationId: getEvents
      security:
        - ApiKeyAuth:
            - events:read
      parameters:
        - name: event_type
          in: query
          description: Only deliver these event types.
          explode: true
          style: form
          schema:
            type: array
            items:
              type: string
        - name: labels
          in: query
          description: JSON object of labels to filter on.
          schema:
            type: string
        - name: token
          in: query
          description: API key, for clients that cannot set an Authorization header.
          schema:
            type: string
      responses:
        "101":
          description: The connection was upgraded to a WebSocket.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the events:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/logs/{taskID}/stream:
    get:
      tags:
        - Streaming
      summary: Stream task logs over WebSocket
      description: Upgrades to a WebSocket that emits one JSON log message per line. Browsers cannot set headers on a WebSocket handshake, so this route also accepts the API key as a ?token= query parameter.
      operationId: getLogsStream
      security:
        - ApiKeyAuth:
            - tasks:read
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          description: API key, for clients that cannot set an Authorization header.
          schema:
            type: string
      responses:
        "101":
          description: The connection was upgraded to a WebSocket.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tasks:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/permissions:
    get:
      tags:
        - Permissions
      summary: List permission requests
      description: 'Permission requests never expire on their own: they stay pending until approved, denied, or cancelled with the task.'
      operationId: getPermissions
      security:
        - ApiKeyAuth:
            - permissions:read
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return. Defaults to 50.
          schema:
            type: integer
            format: int32
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor.
          schema:
            type: string
        - name: session_id
          in: query
          description: Filter by session.
          schema:
            type: string
        - name: task_id
          in: query
          description: Filter by task.
          schema:
            type: string
        - name: status
          in: query
          description: Filter by request status.
          explode: true
          style: form
          schema:
            type: array
            items:
              type: string
        - name: risk_level
          in: query
          description: Filter by risk level.
          explode: true
          style: form
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionRequestList'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the permissions:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/permissions/{permissionID}:
    get:
      tags:
        - Permissions
      summary: Get a permission request
      operationId: getPermissionsByPermissionID
      security:
        - ApiKeyAuth:
            - permissions:read
      parameters:
        - name: permissionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionRequest'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the permissions:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/permissions/{permissionID}/approve:
    post:
      tags:
        - Permissions
      summary: Approve a permission request
      description: May be called while the session is suspended; the response is delivered when the session resumes.
      operationId: postPermissionsApprove
      security:
        - ApiKeyAuth:
            - permissions:write
      parameters:
        - name: permissionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        "204":
          description: Success, with no response body.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the permissions:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/permissions/{permissionID}/deny:
    post:
      tags:
        - Permissions
      summary: Deny a permission request
      operationId: postPermissionsDeny
      security:
        - ApiKeyAuth:
            - permissions:write
      parameters:
        - name: permissionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DenyRequest'
      responses:
        "204":
          description: Success, with no response body.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the permissions:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runners:
    get:
      tags:
        - Runners
      summary: List runners
      description: Runners are managed by the server; this view is read-only. Provisioning lives in the admin API.
      operationId: getRunners
      security:
        - ApiKeyAuth:
            - runners:read
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return. Defaults to 50.
          schema:
            type: integer
            format: int32
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor.
          schema:
            type: string
        - name: status
          in: query
          description: Filter by runner status.
          explode: true
          style: form
          schema:
            type: array
            items:
              type: string
        - name: pool_name
          in: query
          description: Filter by pool.
          schema:
            type: string
        - name: labels[key]
          in: query
          description: Filter by label. Repeat with different keys to AND several labels, e.g. labels[env]=prod.
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerList'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the runners:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runners/{runnerID}:
    get:
      tags:
        - Runners
      summary: Get a runner
      operationId: getRunnersByRunnerID
      security:
        - ApiKeyAuth:
            - runners:read
      parameters:
        - name: runnerID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runner'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the runners:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/scheduled-tasks:
    post:
      tags:
        - Scheduled Tasks
      summary: Create a scheduled task
      operationId: postScheduledTasks
      security:
        - ApiKeyAuth:
            - scheduled-tasks:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduledTaskRequest'
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTask'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Scheduled Tasks
      summary: List scheduled tasks
      operationId: getScheduledTasks
      security:
        - ApiKeyAuth:
            - scheduled-tasks:read
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return. Defaults to 50.
          schema:
            type: integer
            format: int32
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor.
          schema:
            type: string
        - name: session_id
          in: query
          description: Filter by session.
          schema:
            type: string
        - name: status
          in: query
          description: Filter by scheduled task status.
          explode: true
          style: form
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTaskList'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/scheduled-tasks/{scheduledTaskID}:
    get:
      tags:
        - Scheduled Tasks
      summary: Get a scheduled task
      operationId: getScheduledTasksByScheduledTaskID
      security:
        - ApiKeyAuth:
            - scheduled-tasks:read
      parameters:
        - name: scheduledTaskID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTask'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:read scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Scheduled Tasks
      summary: Update a scheduled task
      operationId: patchScheduledTasksByScheduledTaskID
      security:
        - ApiKeyAuth:
            - scheduled-tasks:write
      parameters:
        - name: scheduledTaskID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduledTaskRequest'
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTask'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Scheduled Tasks
      summary: Delete a scheduled task
      operationId: deleteScheduledTasksByScheduledTaskID
      security:
        - ApiKeyAuth:
            - scheduled-tasks:write
      parameters:
        - name: scheduledTaskID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Success, with no response body.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/scheduled-tasks/{scheduledTaskID}/pause:
    post:
      tags:
        - Scheduled Tasks
      summary: Pause a scheduled task
      operationId: postScheduledTasksPause
      security:
        - ApiKeyAuth:
            - scheduled-tasks:write
      parameters:
        - name: scheduledTaskID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTask'
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the scheduled-tasks:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/tasks/{taskID}/cancel:
    post:
      tags:
        - Tasks
      summary: Cancel a task
      operationId: postTasksCancel
      security:
        - ApiKeyAuth:
            - tasks:write
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Success, with no response body.
        "401":
          description: The API key is missing, malformed, revoked or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: The API key lacks the tasks:write scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: An error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      bearerFormat: mk_...
      description: An API key minted by the admin API.
      scheme: bearer
      type: http
  schemas:
    ApproveRequest:
      type: object
      properties:
        reason:
          type: string
...
    DenyRequest:
      type: object
      properties:
        reason:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    PermissionRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PermissionRequest'
        total_count:
          type: integer
          format: int64
        has_more:
          type: boolean
        next_cursor:
          type: string
      required:
        - items
        - total_count
        - has_more
...